Libraries updates and cleanup
 * Updating common/lib
 * Updating lib/csu
 * Updating lib/libc
 * Updating libexec/ld.elf_so
 * Corrected test on __minix in featuretest to actually follow the
   meaning of the comment.
 * Cleaned up _REENTRANT-related defintions.
 * Disabled -D_REENTRANT for libfetch
 * Removing some unneeded __NBSD_LIBC defines and tests

+++ lib/libc/resolv/res_debug.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: res_debug.h,v 1.1.1.4 2009/04/12 16:35:47 christos Exp $	*/
+/*	$NetBSD: res_debug.h,v 1.2 2012/03/13 21:13:43 christos Exp $	*/
 
 /*
  * Copyright (c) 2004 by Internet Systems Consortium, Inc. ("ISC")
@@ -29,7 +29,7 @@
 #   define Dprint(cond, args) if (cond) {fprintf args;} else {}
 #   define DprintQ(cond, args, query, size) if (cond) {\
 			fprintf args;\
-			res_pquery(statp, query, size, stdout);\
+			res_pquery(statp, (query), (int)(size), stdout);\
 		} else {}
 #endif
